UPDATE: m_size type

+++ include/rdma_common_doca.h
@@ -220,7 +220,7 @@ extern "C"
      */
     doca_error_t allocate_rdma_resources(struct rdma_config *cfg, const uint32_t mmap_permissions,
                                          const uint32_t rdma_permissions, tasks_check func,
-                                         struct rdma_resources *resources, uint32_t m_size, uint16_t n_conn);
+                                         struct rdma_resources *resources, uint64_t m_size, uint16_t n_conn);
 
     /*
      * Destroy DOCA RDMA resources
